Update cava-colors
Fixes as an error that does not display cava and gives this error-
```Error loading config. Only 'ncurses' output method supports HTML colors. Please change the colours or the output method.
As of version 0.7.0 ncurses is no longer the default output method
Gradient color 1 is invalid. It must be HTML color of the form '#xxxxxx'.```