Fix: Update wallpaper path in WallustSwww.sh
Updated wallpaper path in WallustSwww.sh to match the newest version of swww. It worked in a test on my local machine on fedora. The change was necessary to let ~/.config/rofi/.current_wallpaper to update automatically and it also fixed wallust updating the color scheme automatically. It also made hyprlock to use the current wallpaper as background. Rofi's embed wallpaper section got fixed too.