Modify SwitchKeyboardLayout.sh to not require .cache/kb_layout
The bulk of the SwitchKeyboardLayout.sh script has been renamed to
KeyboardLayout.sh, meanwhile its behavior has been changed.

Instead of relying on the file .cache/kb_layout, to check the current
keyboard layout, as well as to toggle it, it now uses hyprctl to query
this information, while making sure that it checks the first non-ignored
keyboard. In this way, it querys from keyboards which it's also in
charge of modifying. The logic from the script is also repurposed in
waybar, just for viewing, so the script was renamed.